Fix invalid use of 'this' in math_base.hpp helper functions
The overrideBinaryActivateForType and overrideUnaryActivateForType
template functions used macros containing decltype(this), which is
only valid inside member functions. Replaced the macro-based approach
with clean template lambdas that explicitly use TShard from the
function template parameter.
