+++ justfile
@@ -64,6 +64,9 @@ tests:
   build/Debug/shards shards/tests/general.shs
   build/Debug/shards shards/tests/hello.shs
 
+run-all-tests mode="Release":
+  sh run-macos-gpu-tests.sh --with-cpu --shards-bin build/{{mode}}/shards
+
 format:
   sh format.sh
